Make the certificate installation flow compatible with --deploy-hook
The script can be now used with certbot certonly/run/renew --deploy-hook
"cpanel-dns.py install example.com" to automatically install certs. If
the domain is not provided on the command line, the script will try to
figure out which CPanel domains to configure with the issued cert from
Certbot's RENEWED_DOMAINS - this is not battle-tested, but should work.

The CPANEL_DNS_INSTALL_TARGET_DOMAIN variable is no longer used.